Make autoformat consistent. Rename CliLogin to UserLogin and move it to exabel_data_sdk/client/user_login.py
