+++ exabel_data_sdk/tests/client/api/test_export_api.py
@@ -11,30 +11,28 @@ class TestExportApi(unittest.TestCase):
         mock = MagicMock(name="run_query")
         export_api.run_query = mock
 
-        def check_query(sql: str) -> None:
-            query = mock.call_args.args[0]
-            self.assertEqual(sql, query.sql())
-
         export_api.signal_query(Column("EURUSD", "fx('EUR', 'USD')"))
-        check_query("SELECT time, 'fx(''EUR'', ''USD'')' AS EURUSD FROM signals")
+        mock.assert_called_with("SELECT time, 'fx(''EUR'', ''USD'')' AS EURUSD FROM signals")
 
         export_api.signal_query("Sales_Actual", factset_id="QLGSL2-R")
-        check_query("SELECT time, Sales_Actual FROM signals WHERE factset_id = 'QLGSL2-R'")
+        mock.assert_called_with(
+            "SELECT time, Sales_Actual FROM signals WHERE factset_id = 'QLGSL2-R'"
+        )
 
         export_api.signal_query("Sales_Actual", bloomberg_ticker=["AAPL US", "TSLA US"])
-        check_query(
+        mock.assert_called_with(
             "SELECT name, time, Sales_Actual FROM signals "
             "WHERE bloomberg_ticker IN ('AAPL US', 'TSLA US')"
         )
 
         export_api.signal_query("Sales_Actual", tag="graph:tag:user:xyz-123")
-        check_query(
+        mock.assert_called_with(
             "SELECT name, time, Sales_Actual FROM signals WHERE has_tag('graph:tag:user:xyz-123')"
         )
 
         column = Column("Q", "sales_actual(alignment='afp')")
         export_api.signal_query(["Sales_Actual_fiscal", column], "AAPL US")
-        check_query(
+        mock.assert_called_with(
             "SELECT time, Sales_Actual_fiscal, 'sales_actual(alignment=''afp'')' AS Q "
             "FROM signals WHERE bloomberg_ticker = 'AAPL US'"
         )
